refactor: :recycle: Update dependencies and restructure auth config
This commit updates several dependencies in `package.json` and `package-lock.json` to their latest versions. Additionally, it removes the `auth.config.ts` file and refactors its usage throughout the application. The changes ensure streamlined authentication handling by directly importing from `@/auth` instead of relying on a separate configuration file.

**Breaking Changes or Issues Closed:**
- Removed `auth.config.ts`, affecting modules relying on its exports.
- Resolves compatibility issues with outdated dependencies.
- Delete ProductCloudinaryGallery
