+++ package.json
@@ -17,65 +17,65 @@
     "reset:soft": "npm run clean:soft && npm run seed && clear && npm run dev"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.9.0",
-    "@paypal/react-paypal-js": "^8.1.3",
-    "@prisma/client": "^5.11.0",
-    "@radix-ui/react-accordion": "^1.2.0",
-    "@radix-ui/react-avatar": "^1.1.0",
-    "@radix-ui/react-dialog": "^1.1.2",
-    "@radix-ui/react-dropdown-menu": "^2.1.1",
-    "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-popover": "^1.1.1",
-    "@radix-ui/react-radio-group": "^1.2.1",
-    "@radix-ui/react-select": "^2.1.1",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-switch": "^1.1.0",
-    "@tailwindcss/container-queries": "^0.1.1",
-    "bcryptjs": "^2.4.3",
-    "class-variance-authority": "^0.7.0",
-    "cloudinary": "^2.2.0",
-    "clsx": "^2.1.1",
-    "embla-carousel-autoplay": "^8.1.5",
-    "embla-carousel-react": "^8.1.7",
-    "next": "^14.2.4",
-    "next-auth": "^5.0.0-beta.15",
-    "next-cloudinary": "^6.12.0",
-    "next-themes": "^0.3.0",
-    "nextjs-toploader": "^3.6.15",
-    "nodemailer": "^6.9.14",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-hook-form": "^7.52.2",
-    "react-icons": "^5.0.1",
-    "recharts": "^2.12.7",
-    "rizzui": "^0.8.7",
-    "sonner": "^1.5.0",
-    "tailwind-merge": "^2.4.0",
-    "tailwindcss-animate": "^1.0.7",
-    "twilio": "^5.0.4",
-    "use-debounce": "^10.0.1",
-    "zod": "^3.23.8",
-    "zustand": "^4.5.2"
+    "@hookform/resolvers": "3.9.0",
+    "@paypal/react-paypal-js": "8.1.3",
+    "@prisma/client": "5.11.0",
+    "@radix-ui/react-accordion": "1.2.0",
+    "@radix-ui/react-avatar": "1.1.0",
+    "@radix-ui/react-dialog": "1.1.2",
+    "@radix-ui/react-dropdown-menu": "2.1.1",
+    "@radix-ui/react-label": "2.1.0",
+    "@radix-ui/react-popover": "1.1.1",
+    "@radix-ui/react-radio-group": "1.2.1",
+    "@radix-ui/react-select": "2.1.1",
+    "@radix-ui/react-slot": "1.1.0",
+    "@radix-ui/react-switch": "1.1.0",
+    "@tailwindcss/container-queries": "0.1.1",
+    "bcryptjs": "2.4.3",
+    "class-variance-authority": "0.7.0",
+    "cloudinary": "2.2.0",
+    "clsx": "2.1.1",
+    "embla-carousel-autoplay": "8.1.5",
+    "embla-carousel-react": "8.1.7",
+    "next": "14.2.4",
+    "next-auth": "5.0.0-beta.25",
+    "next-cloudinary": "6.12.0",
+    "next-themes": "0.3.0",
+    "nextjs-toploader": "3.6.15",
+    "nodemailer": "6.9.14",
+    "react": "18",
+    "react-dom": "18",
+    "react-hook-form": "7.52.2",
+    "react-icons": "5.0.1",
+    "recharts": "2.12.7",
+    "rizzui": "0.8.7",
+    "sonner": "1.5.0",
+    "tailwind-merge": "2.4.0",
+    "tailwindcss-animate": "1.0.7",
+    "twilio": "5.0.4",
+    "use-debounce": "10.0.1",
+    "zod": "3.23.8",
+    "zustand": "4.5.2"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.6",
-    "@types/node": "^20",
-    "@types/nodemailer": "^6.4.15",
-    "@types/qrcode-terminal": "^0.12.2",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "@typescript-eslint/eslint-plugin": "^6.21.0",
-    "autoprefixer": "^10.0.1",
-    "eslint": "^8.57.0",
-    "eslint-config-standard-with-typescript": "^43.0.1",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-react": "^7.33.2",
-    "postcss": "^8",
-    "prisma": "^5.11.0",
-    "tailwindcss": "^3.3.0",
-    "ts-node": "^10.9.2",
+    "@types/bcryptjs": "2.4.6",
+    "@types/node": "20",
+    "@types/nodemailer": "6.4.15",
+    "@types/qrcode-terminal": "0.12.2",
+    "@types/react": "18",
+    "@types/react-dom": "18",
+    "@typescript-eslint/eslint-plugin": "6.21.0",
+    "autoprefixer": "10.0.1",
+    "eslint": "8.57.0",
+    "eslint-config-standard-with-typescript": "43.0.1",
+    "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-n": "16.6.2",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-react": "7.33.2",
+    "postcss": "8",
+    "prisma": "5.11.0",
+    "tailwindcss": "3.3.0",
+    "ts-node": "10.9.2",
     "typescript": "5.3.3"
   }
 }
