+++ src/actions/auth/getUserSessionServer.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { auth } from '@/auth.config'
+import { auth } from '@/auth'
 
 export const getUserSessionServer = async () => {
   const session = await auth()
