+++ src/actions/auth/login.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { signIn } from '@/auth.config'
+import { signIn } from '@/auth'
 
 export async function authenticate(
   prevState: string | undefined,
