+++ src/actions/auth/logout.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { signOut } from '@/auth.config'
+import { signOut } from '@/auth'
 
 export const logout = async () => {
   await signOut()
