+++ src/app/(shop)/checkout/layout.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation'
-import { auth } from '@/auth.config'
+import { auth } from '@/auth'
 
 export default async function CheckoutLayout({
   children
